upgrade visual regression testing tests
